Final Project of the course
The main focuses of this project are the PythonCode.py and CS210_Starter_CPP_Code.obj. CS210_Starter_CPP_Code.obj was given to me with the 3 functions focused on python integration already made, everything else is my own making. PythonCode.py was given to me 2ith printsomething, PrintMe, and SquareValue function already made for example. The rest of the file was my doing. 

Using these files as a basis I had to mix together python and c++ to make a functional menu with functional options.